Sprint 14.1

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Request/DemographicsRequest.cs
@@ -3,7 +3,7 @@ namespace Demoulas.ProfitSharing.Common.Contracts.Request;
 
 public record DemographicsRequest : IMemberRequest
 {
-    public required long Ssn { get; set; }
+    public required int Ssn { get; set; }
     public int EmployeeId { get; set; }
 
     public required long OracleHcmId { get; set; }
